+++ cockatrice/src/server/user/user_list_proxy.h
